update: remove gms from denylist
Due to how some integrity modules work, adding gms to the denylist can often result in integrity failures. Osmosis fork of integrityfix does resolve this through scripts but that only happens once after a device reboot. If gms is added back to the denylist, this can result in failures for the module to properly work and render a basic/device verdict with a working fingerprint.